refactor: domain to application module
