+++ sseudam-core/core-api/src/test/kotlin/com/sseudam/presentation/v1/trashspot/TrashSpotControllerTest.kt
@@ -19,9 +19,9 @@ import com.sseudam.docs.support.STRING
 import com.sseudam.docs.support.parameterType
 import com.sseudam.docs.support.type
 import com.sseudam.fixture.trashspot.TrashSpotFixture
-import com.sseudam.trashspot.TrashSpotDetail
 import com.sseudam.trashspot.TrashSpotFacade
 import com.sseudam.trashspot.TrashType
+import com.sseudam.trashspot.result.TrashSpotDetail
 import io.mockk.every
 import io.mockk.mockk
 import io.restassured.http.ContentType
