+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/auth/dto/GrantedAuthority.kt
@@ -1,4 +1,6 @@
-package com.sseudam.auth
+package com.sseudam.auth.dto
+
+import com.sseudam.auth.AuthorityType
 
 data class GrantedAuthority(
     val authorityType: AuthorityType,
