+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/history/dto/HistoryStatus.kt
@@ -1,4 +1,4 @@
-package com.sseudam.history
+package com.sseudam.history.dto
 
 enum class HistoryStatus(
     val displayName: String,
