+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/history/dto/SpotHistory.kt
@@ -1,4 +1,4 @@
-package com.sseudam.history
+package com.sseudam.history.dto
 
 import com.sseudam.common.Address
 import com.sseudam.common.GeoJson
