+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/notification/dto/FcmKey.kt
@@ -1,4 +1,4 @@
-package com.sseudam.notification.fcm
+package com.sseudam.notification.dto
 
 data class FcmKey(
     val key: String,
