+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/notification/dto/FirebaseCloudMessage.kt
@@ -1,4 +1,4 @@
-package com.sseudam.notification.fcm
+package com.sseudam.notification.dto
 
 data class FirebaseCloudMessage(
     val fcmKey: FcmKey,
