+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/notification/repository/FcmRepository.kt
@@ -1,4 +1,6 @@
-package com.sseudam.notification.fcm
+package com.sseudam.notification.repository
+
+import com.sseudam.notification.dto.FirebaseCloudMessage
 
 interface FcmRepository {
     fun sendAll(firebaseCloudMessages: List<FirebaseCloudMessage>): List<FirebaseCloudMessage>
