+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/report/event/SpotReportEventHandler.kt
@@ -1,8 +1,8 @@
 package com.sseudam.report.event
 
-import com.sseudam.notification.NotificationMessages
-import com.sseudam.notification.SendNotificationMessage
 import com.sseudam.notification.discord.DiscordClient
+import com.sseudam.notification.dto.NotificationMessages
+import com.sseudam.notification.dto.SendNotificationMessage
 import com.sseudam.notification.fcm.FcmSender
 import com.sseudam.report.ReportStatus
 import com.sseudam.report.ReportType
