--- sseudam-core/core-domain/src/main/kotlin/com/sseudam/notification/fcm/FcmToUserKey.kt
@@ -1,5 +1,0 @@
-package com.sseudam.notification.fcm
-
-data class FcmToUserKey(
-    val key: String,
-)
